pyproject.toml
[tool.poetry]
name = "quizgpt"
version = "0.1.0"
description = "Instantly create quizzes on every subject under the sun with the power of Chat-GPT."
authors = ["John Sudds <jsuddsjr@users.noreply.github.com>"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Web Environment",
    "Framework :: Django :: 4",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.11",
    "Topic :: Education :: Computer Aided Instruction (CAI)",
]
license = "MIT"
readme = "README.md"
packages = [{ include = "quizgpt" }]

[tool.poetry.dependencies]
black                = "^23.9.1"
Django               = "^4.2.5"
mypy                 = "^1.5.1"
openai               = "^0.28.0"
poetry               = "^1.6.1"
pytest               = "^7.4.2"
pytest-django        = "^4.5.2"
python               = "^3.11"
python-dotenv        = "^1.0.0"
django-debug-toolbar = "^4.2.0"
django-ninja         = "^1.0b1"

[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
